requirements.txt
Django>=3.2.9,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.7.5,<2.8.0
pyjwt>=2.3.0,<2.4.0
django-extensions>=3.1.0,<3.6.0
django-cors-headers==3.5.0
djangorestframework-simplejwt==5.0.0
flake8>=3.6.0,<3.7.0